requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==1.2.1
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llama-index-core==0.14.10
llama-index-embeddings-gemini==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-instrumentation==0.4.2
llama-index-llms-gemini==0.6.1
llama-index-llms-google-genai==0.6.0
llama-index-llms-openai==0.6.10
llama-index-multi-modal-llms-gemini==0.6.1
llama-index-readers-file==0.5.5
llama-index-workflows==2.11.5
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
numpy==2.2.6
openai==2.9.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.5.1
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pymupdf==1.26.6
pyparsing==3.2.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
streamlit==1.52.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.235
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.0
watchdog==6.0.0
wrapt==2.0.1
yarl==1.22.0
